mall + trade：获得订单结算信息，接入支付接口

+++ yudao-module-member/yudao-module-member-api/src/main/java/cn/iocoder/yudao/module/member/api/address/dto/AddressRespDTO.java
@@ -29,7 +29,7 @@ public class AddressRespDTO {
     /**
      * 地区编号
      */
-    private Long areaId;
+    private Integer areaId;
     /**
      * 邮编
      */
